+++ CHANGELOG.md
@@ -7,10 +7,20 @@
  ## [Unreleased]
 
 ## [0.0.2] - 2025-06-11
+
 ### Added
-- Integrated `black` code formatter to ensure consistent code style across the project.
+ - Integrated `black` code formatter to ensure consistent code style across the project.
+ - Added `remove_local_folder` task in `upload_to_azure_storage_dag` to remove local files. 
+
+### Changed 
+ - `upload_to_azure_storage_dag` to upload multiple files on azure storage.
+ -  Changed name of `upload_to_azure_storage_dag` to `save_and_delete`.
+ 
+### Fixed
+ -  `trigger_upload_to_azure_storage_dag` by removing hardcoded `channel_id`.
 
 ## [0.0.1] - 2025-06-11
+
 ### Added 
  - Initializes `CHANGELOG.md` to track project changes. 
  - Added `apache-airflow-providers-microsoft-azure` lib to work with azure storage.